// src/admin/features/common/enums.ts
export enum EPages {
  AUTHORIZATION = "AUTHORIZATION",
  COMPANY = "COMPANY",
  ORDERS = "ORDERS",
  TABLES = "TABLES",
  ELEMENTS = "ELEMENTS",
  CATEGORIES = "CATEGORIES",
  KITCHEN = "KITCHEN",
  POSITIONS = "POSITIONS",
  USERS = "USERS",
  CASH_REPORT_MAKE = "CASH_REPORT_MAKE",
  CASH_REPORT_HISTORY = "CASH_REPORT_HISTORY",
  STATS = "STATS",
  BILLING = "BILLING",
  DAY_STATS = "DAY_STATS",
  PERIOD = "PERIOD",
}